Migrate db set medium int to int

+++ db/migrations/20160204130924_CreateTableUsers.sql
@@ -6,7 +6,7 @@ CREATE TABLE `users` (
   `email` VARCHAR(255) NOT NULL COMMENT 'unique email',
   `bitcoin_address` VARCHAR(63) NOT NULL COMMENT 'unique bitcoin address',
   `status` VARCHAR(15) NOT NULL DEFAULT 'unverified' COMMENT 'indicate account status, can be unverified|verified|banned',
-  `balance` MEDIUMINT(8) NOT NULL DEFAULT 0 COMMENT 'user account balance count in bitcoin satonish',
+  `balance` INT(11) NOT NULL DEFAULT 0 COMMENT 'user account balance count in bitcoin satonish',
   `min_withdrawal_amount` MEDIUMINT(8) NOT NULL DEFAULT 100000 COMMENT 'minimum withdrawal amount',
   `reward_interval` SMALLINT(6) NOT NULL DEFAULT 900 COMMENT 'users can get reward every $reward_interval seconds',
   `rewarded_at` DATETIME NOT NULL DEFAULT '1970-01-01 00:00:01',
